REF: Cleared out multiple footers
