+++ src/components/elements/Footer/Footer.jsx
@@ -1,36 +1,84 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import Grid from '@material-ui/core/Grid'
+import { makeStyles } from '@material-ui/core/styles'
+import Container from '@material-ui/core/Container'
 
-import MicroFooter from './MicroFooter'
-import SmallFooter from './SmallFooter'
-import BigFooter from './BigFooter'
+import Copyright from '../Copyright'
+import GradientStripe from '../GradientStripe'
+import { Link } from '../../core/Link'
 
-/**
- * Footer
- *
- * A dark themed footer for the site. Provides multiple different kinds.
- */
-function Footer({ kind, ...rest }) {
-  switch (kind) {
-    case 'micro':
-      return <MicroFooter {...rest} />
-    case 'small':
-      return <SmallFooter {...rest} />
-    case 'big':
-      return <BigFooter {...rest} />
-    default:
-      return null
-  }
+import { FaTwitter, FaLinkedin, FaGithub } from 'react-icons/fa'
+
+const useStyles = makeStyles((theme) => ({
+  footer: {
+    backgroundColor: theme.palette.background.paper,
+  },
+  button: {
+    margin: theme.spacing(1),
+    fontSize: '24px',
+  },
+  copyright: {
+    fontSize: `${theme.spacing(1.5)}px`,
+  },
+}))
+
+function SmallFooter({ id }) {
+  const classes = useStyles()
+  return (
+    <footer id={id} className={classes.footer}>
+      <GradientStripe />
+      <Container>
+        <Grid
+          container
+          direction="row"
+          justify="space-between"
+          alignItems="center"
+        >
+          <Grid item>
+            <Copyright className={classes.copyright} />
+          </Grid>
+          <Grid item>
+            <Link
+              href="https://twitter.com/octue"
+              kind="external"
+              componentType="iconButton"
+              className={classes.button}
+              aria-label="twitter"
+            >
+              <FaTwitter />
+            </Link>
+            <Link
+              href="https://linkedin.com/company/octue"
+              kind="external"
+              componentType="iconButton"
+              className={classes.button}
+              aria-label="linkedin"
+            >
+              <FaLinkedin />
+            </Link>
+            <Link
+              href="https://github.com/octue"
+              kind="external"
+              componentType="iconButton"
+              className={classes.button}
+              aria-label="github"
+            >
+              <FaGithub />
+            </Link>
+          </Grid>
+        </Grid>
+      </Container>
+    </footer>
+  )
 }
 
-Footer.defaultProps = {
+SmallFooter.defaultProps = {
   id: 'footer',
-  kind: 'small',
 }
 
-Footer.propTypes = {
-  kind: PropTypes.oneOf(['big', 'small', 'micro', 'none']),
+SmallFooter.propTypes = {
   id: PropTypes.string,
 }
 
-export default Footer
+export default SmallFooter
