+++ src/components/elements/Footer/index.js
@@ -1,4 +1,1 @@
 export { default } from './Footer'
-export { default as MicroFooter } from './MicroFooter'
-export { default as SmallFooter } from './SmallFooter'
-export { default as BigFooter } from './BigFooter'
