+++ src/components/elements/index.js
@@ -2,5 +2,6 @@ export { default as CallToAction } from './CallToAction'
 export { default as CookieBar } from './CookieBar'
 export { default as Copyright } from './Copyright'
 export { default as GradientStripe } from './GradientStripe'
+export { default as Footer } from './Footer'
 export { default as ProTip } from './ProTip'
 export { default as TwitterEmbed } from './TwitterEmbed'
