+++ src/components/layout/index.js
@@ -1,7 +1,6 @@
 export { default as BackgroundBox } from './BackgroundBox'
 export { default as CardPage } from './CardPage'
 export { default as ErrorPage } from './ErrorPage'
-export { default as Footer } from './Footer'
 export { default as IFrameModal } from './IFrameModal'
 export { default as Main } from './Main'
 export { default as NavBar, NavItem, NavItems } from './NavBar'
