+++ src/containers/layout/BasicPage/BasicPage.jsx
@@ -2,7 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import Box from '@material-ui/core/Box'
 import { ThemeProvider } from '@material-ui/styles'
-import { NavBar, NavItems, Main, Footer } from '../../../components/layout'
+import { Footer } from '../../../components/elements'
+import { NavBar, NavItems, Main } from '../../../components/layout'
 import { lightTheme, darkTheme } from '../../../themes'
 import Toastable from '../Toastable'
 import CookieBar from '../../../components/elements/CookieBar'
